Added the centre() method and made a few tweaks
Have added the centre() method which can be used in conjunction with the
indent() method or puts() to centre text. Have updated the indent method
as I didn't understand how it worked and changed the default "spaces"
value to 30 as I felt the 44 was too wide for the screen.
I've also saved the README as a .md file as that's the format typically
used in repos.